[Secure Config] Refactored SafetyRulesBackend into SecureBackend and created a SecureConfig as a template to hold all Secure crate specific related configurations.
