+++ config/src/config/mod.rs
@@ -32,6 +32,8 @@ mod mempool_config;
 pub use mempool_config::*;
 mod network_config;
 pub use network_config::*;
+mod secure_config;
+pub use secure_config::*;
 mod state_sync_config;
 pub use state_sync_config::*;
 mod storage_config;
@@ -72,6 +74,8 @@ pub struct NodeConfig {
     pub metrics: MetricsConfig,
     #[serde(default)]
     pub mempool: MempoolConfig,
+    #[serde(default)]
+    pub secure: SecureConfig,
     #[serde(default)]
     pub state_sync: StateSyncConfig,
     #[serde(default)]
@@ -175,6 +179,7 @@ impl NodeConfig {
             metrics: self.metrics.clone(),
             mempool: self.mempool.clone(),
             state_sync: self.state_sync.clone(),
+            secure: self.secure.clone(),
             storage: self.storage.clone(),
             test: None,
             upstream: self.upstream.clone(),
