+++ config/src/config/safety_rules_config.rs
@@ -1,20 +1,21 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::config::SecureBackend;
 use serde::{Deserialize, Serialize};
 use std::{net::SocketAddr, path::PathBuf};
 
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 #[serde(default, deny_unknown_fields)]
 pub struct SafetyRulesConfig {
-    pub backend: SafetyRulesBackend,
+    pub backend: SecureBackend,
     pub service: SafetyRulesService,
 }
 
 impl Default for SafetyRulesConfig {
     fn default() -> Self {
         Self {
-            backend: SafetyRulesBackend::InMemoryStorage,
+            backend: SecureBackend::InMemoryStorage,
             service: SafetyRulesService::Local,
         }
     }
@@ -22,69 +23,10 @@ impl Default for SafetyRulesConfig {
 
 impl SafetyRulesConfig {
     pub fn set_data_dir(&mut self, data_dir: PathBuf) {
-        if let SafetyRulesBackend::OnDiskStorage(backend) = &mut self.backend {
+        if let SecureBackend::OnDiskStorage(backend) = &mut self.backend {
             backend.set_data_dir(data_dir);
         }
     }
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(tag = "type")]
-#[serde(rename_all = "snake_case")]
-pub enum SafetyRulesBackend {
-    InMemoryStorage,
-    Vault(VaultConfig),
-    OnDiskStorage(OnDiskStorageConfig),
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-pub struct OnDiskStorageConfig {
-    // In testing scenarios this implies that the default state is okay if
-    // a state is not specified.
-    pub default: bool,
-    // Required path for on disk storage
-    pub path: PathBuf,
-    #[serde(skip)]
-    data_dir: PathBuf,
-}
-
-impl Default for OnDiskStorageConfig {
-    fn default() -> Self {
-        Self {
-            default: false,
-            path: PathBuf::from("safety_rules.toml"),
-            data_dir: PathBuf::from("/opt/libra/data/common"),
-        }
-    }
-}
-
-impl OnDiskStorageConfig {
-    pub fn path(&self) -> PathBuf {
-        if self.path.is_relative() {
-            self.data_dir.join(&self.path)
-        } else {
-            self.path.clone()
-        }
-    }
-
-    pub fn set_data_dir(&mut self, data_dir: PathBuf) {
-        self.data_dir = data_dir;
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-pub struct VaultConfig {
-    /// In testing scenarios this will install baseline data if it is not specified. Note: this can
-    /// only be used if the token provided has root or sudo access.
-    pub default: bool,
-    /// A namespace is an optional portion of the path to a key stored within Vault. For example,
-    /// a secret, S, without a namespace would be available in secret/data/S, with a namespace, N, it
-    /// would be in secret/data/N/S.
-    pub namespace: Option<String>,
-    /// Vault's URL, note: only HTTP is currently supported.
-    pub server: String,
-    /// The authorization token for access secrets
-    pub token: String,
 }
 
 /// Defines how safety rules should be executed
