+++ config/src/config/secure_config.rs
@@ -1,0 +1,74 @@
+// Copyright (c) The Libra Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+use serde::{Deserialize, Serialize};
+use std::path::PathBuf;
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+#[serde(default, deny_unknown_fields)]
+pub struct SecureConfig {}
+
+impl Default for SecureConfig {
+    fn default() -> Self {
+        Self {}
+    }
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+#[serde(tag = "type")]
+#[serde(rename_all = "snake_case")]
+pub enum SecureBackend {
+    InMemoryStorage,
+    Vault(VaultConfig),
+    OnDiskStorage(OnDiskStorageConfig),
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+pub struct VaultConfig {
+    /// In testing scenarios this will install baseline data if it is not specified. Note: this can
+    /// only be used if the token provided has root or sudo access.
+    pub default: bool,
+    /// A namespace is an optional portion of the path to a key stored within Vault. For example,
+    /// a secret, S, without a namespace would be available in secret/data/S, with a namespace, N, it
+    /// would be in secret/data/N/S.
+    pub namespace: Option<String>,
+    /// Vault's URL, note: only HTTP is currently supported.
+    pub server: String,
+    /// The authorization token for access secrets
+    pub token: String,
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+pub struct OnDiskStorageConfig {
+    // In testing scenarios this implies that the default state is okay if
+    // a state is not specified.
+    pub default: bool,
+    // Required path for on disk storage
+    pub path: PathBuf,
+    #[serde(skip)]
+    data_dir: PathBuf,
+}
+
+impl Default for OnDiskStorageConfig {
+    fn default() -> Self {
+        Self {
+            default: false,
+            path: PathBuf::from("safety_rules.toml"),
+            data_dir: PathBuf::from("/opt/libra/data/common"),
+        }
+    }
+}
+
+impl OnDiskStorageConfig {
+    pub fn path(&self) -> PathBuf {
+        if self.path.is_relative() {
+            self.data_dir.join(&self.path)
+        } else {
+            self.path.clone()
+        }
+    }
+
+    pub fn set_data_dir(&mut self, data_dir: PathBuf) {
+        self.data_dir = data_dir;
+    }
+}
