+++ consensus/safety-rules/src/process_client_wrapper.rs
@@ -12,7 +12,7 @@ use consensus_types::{
     vote_proposal::VoteProposal,
 };
 use libra_config::{
-    config::{ConsensusType, NodeConfig, RemoteService, SafetyRulesBackend, SafetyRulesService},
+    config::{ConsensusType, NodeConfig, RemoteService, SafetyRulesService, SecureBackend},
     utils,
 };
 use libra_crypto::ed25519::Ed25519Signature;
@@ -32,7 +32,7 @@ pub struct ProcessClientWrapper<T> {
 }
 
 impl<T: Payload> ProcessClientWrapper<T> {
-    pub fn new(backend: SafetyRulesBackend) -> Self {
+    pub fn new(backend: SecureBackend) -> Self {
         let server_port = utils::get_available_port();
         let server_address = SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), server_port);
